+++ bootstrap/server.go
@@ -30,6 +30,7 @@ func Commands() *mux.Router {
 	rtr.HandleFunc("/status", APIHealth).Methods("GET")
 	rtr.HandleFunc("/api/k3s/v0/config", APIGetKubeConfig).Methods("GET")
 	rtr.HandleFunc("/api/k3s/v0/version", APIGetKubeVersion).Methods("GET")
+	rtr.HandleFunc("/status?verbose", APIStatus).Methods("GET")
 
 	return rtr
 }
@@ -138,6 +139,25 @@ func deployTraefikDashboard() {
 	TraefikDashboardInstalled = true
 }
 
+// APIStatus give out the status of the kubernetes server
+func APIStatus(w http.ResponseWriter, r *http.Request) {
+	w.Header().Set("Content-Type", "text/plain; charset=utf-8")
+	w.Header().Set("Api-Service", "v0")
+
+	logrus.Debug("Status Information")
+
+	// check if the kubernetes server is working
+	stdout, err := exec.Command("/mnt/mesos/sandbox/kubectl", "get", "--raw='/readyz?verbose'").Output()
+
+	if err != nil {
+		logrus.Error("Health to Kubernetes Server: ", err)
+		w.WriteHeader(http.StatusInternalServerError)
+		return
+	}
+
+	w.Write(stdout)
+}
+
 func main() {
 	util.SetLogging("INFO", false, "GO-K3S-API")
 
@@ -153,5 +173,4 @@ func main() {
 	if err := http.ListenAndServe(*bind+":"+*port, nil); err != nil {
 		logrus.Fatalln("ListenAndServe: ", err)
 	}
-
 }